+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix the permission issue during data migration
+- Update bundled resources when version changes
 
 ## [0.2.0](https://github.com/raymond-u/lungo/compare/v0.1.12...v0.2.0) - 2023-11-26