+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Nightly](https://github.com/raymond-u/lungo/compare/v0.2.1...HEAD)
+## [Nightly](https://github.com/raymond-u/lungo/compare/v0.2.2...HEAD)
+
+## [0.2.2](https://github.com/raymond-u/lungo/compare/v0.2.1...v0.2.2) - 2023-11-27
 
 ### Changed